requirements.txt
transformers
Pillow
datasets
soundfile
torch
flask
packaging